Вносит правки по разделению кода на модули
