+++ js/data.js
@@ -1,6 +1,4 @@
-import {getLengthString,getPhotosData} from './util.js';
-
-const SIMILAR_ARRAY_COUNT = 26
+import {getLengthString, getPhotosData} from './util.js';
 
 const COMMENT_LENGTH = 140;
 
@@ -8,4 +6,4 @@ const commentString = 'Функция для проверки максималь
 
 getLengthString(commentString, COMMENT_LENGTH);
 
-getPhotosData(SIMILAR_ARRAY_COUNT);
+export {getPhotosData}
